requirements.txt
numpy
pandas
matplotlib
seaborn
jupyter
tqdm
scipy
python-box
ensure
scikit-learn